chore(Select): convert Select and SelectField to TS
