+++ packages/forma-36-react-components/src/components/CopyButton/CopyButton.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
 import React, { Component } from 'react';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import cn from 'classnames';
