+++ packages/forma-36-react-components/src/components/List/ListItem/ListItem.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
 import React, { Component } from 'react';
 import cn from 'classnames';
