--- packages/forma-36-react-components/src/components/Select/Option/Option.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export class Option extends React.Component {
-  static propTypes = {
-    value: PropTypes.string.isRequired,
-    children: PropTypes.node.isRequired,
-  };
-
-  render() {
-    const { value, children, ...otherProps } = this.props;
-
-    return (
-      <option value={value} {...otherProps}>
-        {children}
-      </option>
-    );
-  }
-}
-
-export default Option;
