+++ packages/forma-36-react-components/src/components/Select/Option/Option.tsx
@@ -1,0 +1,20 @@
+import React, { Component } from 'react';
+
+export interface OptionProps {
+  value: string;
+  children: React.ReactNode;
+}
+
+export class Option extends Component<OptionProps> {
+  render() {
+    const { value, children, ...otherProps } = this.props;
+
+    return (
+      <option value={value} {...otherProps}>
+        {children}
+      </option>
+    );
+  }
+}
+
+export default Option;
