+++ packages/forma-36-react-components/src/components/Select/Select/Select.stories.tsx
@@ -2,13 +2,14 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text, boolean, select } from '@storybook/addon-knobs';
-import { withInfo } from '@storybook/addon-info';
 
 import Select from './Select';
 import Option from '../Option';
 
 storiesOf('Components|Select', module)
-  .addDecorator(withInfo)
+  .addParameters({
+    propTypes: [Select['__docgenInfo'], Option['__docgenInfo']],
+  })
   .add('default', () => (
     <Select
       extraClassNames={text('Extra Class Names', '')}
