+++ packages/forma-36-react-components/src/components/SelectField/SelectField.tsx
@@ -1,44 +1,37 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import cn from 'classnames';
 import ValidationMessage from '../ValidationMessage';
-import FormLabel from '../FormLabel';
+import FormLabel, { FormLabelProps } from '../FormLabel/FormLabel';
 import HelpText from '../HelpText';
-import Select from '../Select/Select';
-import TextLink from '../TextLink';
+import Select, { SelectProps } from '../Select/Select/Select';
+import TextLink, { TextLinkProps } from '../TextLink/TextLink';
 import styles from './SelectField.css';
 
-export class SelectField extends React.Component {
-  static propTypes = {
-    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-    validationMessage: PropTypes.string,
-    testId: PropTypes.string,
-    children: PropTypes.node.isRequired,
-    extraClassNames: PropTypes.string,
-    formLabelProps: PropTypes.object,
-    textLinkProps: PropTypes.object,
-    selectProps: PropTypes.object,
-    name: PropTypes.string.isRequired,
-    id: PropTypes.string.isRequired,
-    labelText: PropTypes.string,
-    helpText: PropTypes.string,
-    required: PropTypes.bool,
-    onChange: PropTypes.func,
-    onBlur: PropTypes.func,
-  };
+export interface SelectFieldProps {
+  name: string;
+  id: string;
+  children: React.ReactNode;
+  value?: string;
+  validationMessage?: string;
+  testId?: string;
+  extraClassNames?: string;
+  formLabelProps?: Partial<FormLabelProps>;
+  textLinkProps?: Partial<TextLinkProps>;
+  selectProps?: Partial<SelectProps>;
+  labelText?: string;
+  helpText?: string;
+  required?: boolean;
+  onChange?: (...args: any[]) => any;
+  onBlur?: (...args: any[]) => any;
+}
 
+export interface SelectFieldState {
+  value?: string;
+}
+
+export class SelectField extends Component<SelectFieldProps, SelectFieldState> {
   static defaultProps = {
-    validationMessage: undefined,
-    extraClassNames: undefined,
     testId: 'cf-ui-select-field',
-    formLabelProps: undefined,
-    textLinkProps: undefined,
-    selectProps: undefined,
-    labelText: undefined,
-    onChange: undefined,
-    helpText: undefined,
-    onBlur: undefined,
-    value: undefined,
     required: false,
   };
 
@@ -82,7 +75,7 @@ export class SelectField extends React.Component {
       ...otherProps
     } = this.props;
 
-    const classNames = cn(styles.SelectField, extraClassNames);
+    const classNames = cn(styles['SelectField'], extraClassNames);
 
     return (
       <div className={classNames} {...otherProps} data-test-id={testId}>
@@ -123,7 +116,7 @@ export class SelectField extends React.Component {
           </ValidationMessage>
         )}
         {helpText && (
-          <div className={styles.SelectField__hints}>
+          <div className={styles['SelectField__hints']}>
             {helpText && (
               <HelpText extraClassNames={styles['SelectField__help-text']}>
                 {helpText}
