+++ packages/forma-36-react-components/src/components/Table/TableCell/TableCell/TableCell.tsx
@@ -35,7 +35,6 @@ export class TableCell extends Component<TableCellProps> {
     return (
       <TableCellContext.Consumer>
         {({ name: context, element, offsetTop }) => {
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
           const Element = element as any;
           return (
             <Element
